oidc: Add helper to get IDP metadata
